+++ packages/admin-panel/src/widgets/index.js
@@ -18,3 +18,4 @@ export {
 export { JsonEditor, JsonTreeEditor } from './JsonEditor';
 export { SecondaryNavbar } from '../layout/navigation/SecondaryNavbar';
 export { ConfirmDeleteModal } from './ConfirmDeleteModal';
+export { EntityOptionLabel } from './EntityOptionLabel';
